refactor: [#661] move Tracker Checker mod
