+++ src/console/clients/checker/console.rs
