+++ src/console/clients/checker/printer.rs
