+++ src/console/clients/mod.rs
@@ -1,1 +1,2 @@
 //! Console clients.
+pub mod checker;
